+++ db2/history2/transaction_q.go
@@ -102,3 +102,8 @@ func (q TransactionsQ) Select() ([]Transaction, error) {
 
 	return result, nil
 }
+
+func (q *TransactionsQ) GetByHash(hash string) (*Transaction, error) {
+	q.selector = q.selector.Where("transactions.hash = ?", hash)
+	return q.Get()
+}