+++ init_web.go
@@ -120,6 +120,7 @@ func initWebActions(app *App) {
 
 	r := app.web.router
 	r.Get("/", &RootAction{})
+
 	// system summary variables too verbose to be included into /
 	r.Get("/statistics", &StatisticsAction{})
 	r.Get("/statistics/balances", &BalancesReportAction{})
@@ -391,15 +392,6 @@ func initWebActions(app *App) {
 
 		// checking if request is signed and deciding on proper handler
 		// (we rely on SignatureValidator middleware here)
-		signer := r.Header.Get(signcontrol.PublicKeyHeader)
-		if signer != "" || app.config.DisableAPISubmit {
-			TransactionCreateAction{}.ServeHTTPC(c, w, r)
-		} else {
-			apiProxy.ServeHTTP(w, r)
-		}
-	}))
-
-	r.Post("/v3/transactions", web.HandlerFunc(func(c web.C, w http.ResponseWriter, r *http.Request) {
 		signer := r.Header.Get(signcontrol.PublicKeyHeader)
 		if signer != "" || app.config.DisableAPISubmit {
 			TransactionCreateAction{}.ServeHTTPC(c, w, r)