+++ init_web_v2.go
@@ -3,6 +3,8 @@ package horizon
 import (
 	"time"
 
+	"gitlab.com/tokend/horizon/corer"
+
 	"gitlab.com/tokend/horizon/web_v2/handlers"
 
 	"net/http"
@@ -57,7 +59,16 @@ func initWebV2Middleware(app *App) {
 			// log will be set by logger setter on handler call
 			ctx.SetHistoryRepo(app.HistoryRepoLogged(nil)),
 			ctx.SetDoorman(doorman.New(app.config.SkipCheck, signersProvider)),
-			ctx.SetCoreInfo(*app.CoreInfo),
+			ctx.SetCoreInfo(func() corer.Info {
+				// required to make sure that we return most recent core info
+				if app.CoreInfo == nil {
+					panic("unexpected state: core info is nil, while we are preparing context for the request")
+				}
+
+				return *app.CoreInfo
+			}),
+			ctx.SetSubmitter(app.submitterV2),
+			ctx.SetConfig(&app.config),
 		),
 		v2middleware.WebMetrics(app),
 	)
@@ -91,6 +102,9 @@ func initWebV2Middleware(app *App) {
 
 func initWebV2Actions(app *App) {
 	m := app.webV2.mux
+
+	m.Get("/v3", handlers.GetRoot)
+	m.Post("/v3/transactions", handlers.CreateTransaction)
 
 	m.Get("/v3/accounts/{id}", handlers.GetAccount)
 	m.Get("/v3/accounts/{id}/signers", handlers.GetAccountSigners)
@@ -178,7 +192,7 @@ func init() {
 	appInit.Add(
 		"web2.init",
 		initWebV2,
-		"app-context", "core-info", "memory_cache", "ledger-state",
+		"app-context", "core-info", "memory_cache", "ledger-state", "submitter_v2",
 	)
 
 	appInit.Add(